+++ packages/eas-cli/src/project/__tests__/projectUtils-test.ts
@@ -1,17 +1,24 @@
 import { vol } from 'memfs';
 
 import { asMock } from '../../__tests__/utils';
+import { confirmAsync } from '../../prompts';
 import { Actor, getUserAsync } from '../../user/User';
+import {
+  ensureProjectExistsAsync,
+  findProjectIdByAccountNameAndSlugNullableAsync,
+} from '../ensureProjectExists';
 import {
   findProjectRootAsync,
   getProjectAccountName,
   getProjectAccountNameAsync,
+  promptToCreateProjectIfNotExistsAsync,
 } from '../projectUtils';
 
 jest.mock('@expo/config');
 jest.mock('fs');
 
 jest.mock('../../user/User');
+jest.mock('../ensureProjectExists');
 
 const originalConsoleLog = console.log;
 const originalConsoleWarn = console.warn;
@@ -166,5 +173,64 @@ describe(getProjectAccountNameAsync, () => {
     await expect(getProjectAccountNameAsync(expWithoutOwner)).rejects.toThrow(
       'manifest property is required'
     );
+  });
+});
+
+describe(promptToCreateProjectIfNotExistsAsync, () => {
+  beforeEach(() => {
+    asMock(getUserAsync).mockReset();
+    asMock(findProjectIdByAccountNameAndSlugNullableAsync).mockReset();
+  });
+
+  it('returns the existing project', async () => {
+    asMock(getUserAsync).mockImplementation((): Actor | undefined => ({
+      __typename: 'User',
+      id: 'user_id',
+      username: 'quin',
+      accounts: [{ id: 'account_id_1', name: 'quin' }],
+      isExpoAdmin: false,
+    }));
+    asMock(findProjectIdByAccountNameAndSlugNullableAsync).mockImplementation(
+      () => 'already-existing-project-id'
+    );
+    const exp: any = {};
+    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
+    expect(projectId).toBe('already-existing-project-id');
+  });
+  it('makes a new project if none exists', async () => {
+    jest.mock('../../prompts');
+    asMock(getUserAsync).mockImplementation((): Actor | undefined => ({
+      __typename: 'User',
+      id: 'user_id',
+      username: 'quin',
+      accounts: [{ id: 'account_id_1', name: 'quin' }],
+      isExpoAdmin: false,
+    }));
+    asMock(findProjectIdByAccountNameAndSlugNullableAsync).mockImplementation(() => null);
+    asMock(ensureProjectExistsAsync).mockImplementation(() => 'new-project-id');
+
+    // user agrees to make new project
+    (confirmAsync as jest.Mock).mockImplementation(() => true);
+    const exp: any = {};
+    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
+    expect(projectId).toBe('new-project-id');
+  });
+  it('aborts making a new project if the user declines', async () => {
+    jest.mock('../../prompts');
+    asMock(getUserAsync).mockImplementation((): Actor | undefined => ({
+      __typename: 'User',
+      id: 'user_id',
+      username: 'quin',
+      accounts: [{ id: 'account_id_1', name: 'quin' }],
+      isExpoAdmin: false,
+    }));
+    asMock(findProjectIdByAccountNameAndSlugNullableAsync).mockImplementation(() => null);
+    asMock(ensureProjectExistsAsync).mockImplementation(() => 'new-project-id');
+
+    // user doesnt agree to make new project
+    (confirmAsync as jest.Mock).mockImplementation(() => false);
+    const exp: any = {};
+    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
+    expect(projectId).toBe(null);
   });
 });